% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxStatement.R
\name{excel_coord_to_numeric}
\alias{excel_coord_to_numeric}
\title{将数据左边进行处理}
\usage{
excel_coord_to_numeric(coord)
}
\arguments{
\item{coord}{提供坐标}
}
\value{
返回值
}
\description{
将数据左边进行处理
}
\examples{
excel_coord_to_numeric()
}
