% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxStatement.R
\name{excel_TaxStatemen_to_num}
\alias{excel_TaxStatemen_to_num}
\title{将excel列转}
\usage{
excel_TaxStatemen_to_num(col_str)
}
\arguments{
\item{col_str}{列名}
}
\value{
返回值
}
\description{
将excel列转
}
\examples{
excel_TaxStatemen_to_num()
}
