% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxStatement.R
\name{TaxStatement_excel}
\alias{TaxStatement_excel}
\title{将汇总表数据写入EXCEL}
\usage{
TaxStatement_excel(
  erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039",
  FYear,
  FMonth,
  FOrgNumber,
  outputDir
)
}
\arguments{
\item{erpToken}{ERP口令}

\item{outputDir}{}
}
\value{
返回值
}
\description{
将汇总表数据写入EXCEL
}
\examples{
TaxStatement_excel()
}
