% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxStatement.R
\name{TaxStatement_meta}
\alias{TaxStatement_meta}
\title{返回元数据信息}
\usage{
TaxStatement_meta(erpToken = "C0426D23-1927-4314-8736-A74B2EF7A039")
}
\arguments{
\item{erpToken}{ERP口令}
}
\value{
返回值
}
\description{
返回元数据信息
}
\examples{
TaxStatement_meta()
}
